interface transmit and receive works, tested against stm32l0. uart_interface class is very hacky, cleanup needed

+++ libprotoserial/interface/interface_id.hpp
@@ -63,7 +63,7 @@ namespace sp
 #ifndef SP_NO_IOSTREAM
 std::ostream& operator<<(std::ostream& os, const sp::interface_identifier & iid) 
 {
-    os << iid.identifier << '-' << iid.instance;
+    os << (int)iid.identifier << '-' << (int)iid.instance;
     return os;
 }
 #endif
